Calls have associated audio data.
Audio data can be uploaded alongside the call metadata when creating
new calls (which will really only be used by the radio backend, of
course).

I need to change the way the files are stored so that it uses
year/month/day/hour or something, so that the archive of calls can be
used independently of the rails app.

I need to define a processor that can take IMBE data and convert it to
various other formats, so that the audio player can select the format
it prefers; for now I'm just testing with mp3s. Oh, don't forget to
update the audio controller to serve up the audio data rather than
html/json describing it.